src/server/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-rest-knox==5.0.2
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
